bug fix for collect64.c
This is the same issue with collec32.c Undefined behavior occurs for PEs that enter the shmem_collect64() that are not part of the Active Set, so we exclude them from entering the call with a branch.  Additionally, their destination output values are somewhat irrelevant, so they shouldn't call show_dst() either.